requirements.txt
# Polymarket SDK
py-clob-client>=0.18.0

# AI Analysis
anthropic>=0.40.0

# Configuration
python-dotenv>=1.0.0

# Async / Networking
websockets>=12.0
aiohttp>=3.9.0

# CLI Output
rich>=13.7.0

# Blockchain
web3>=6.15.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
